Modularize app js
