--- src/sum.js
@@ -1,3 +1,0 @@
-const sum = (a, b) => a + b;
-
-export default sum;
