--- src/sum.test.js
@@ -1,3 +1,0 @@
-import sum from './sum';
-
-it('adds 2 and 3 to be 5', () => expect(sum(2, 3)).toBe(5));
